+++ components/Layout.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import styles from "./layout.module.css";
+import styles from "../styles/Layout.module.css";
 
 type LayoutProps = {
   children: ReactNode;
