+++ components/SortAndFilterDropdowns.tsx
@@ -1,7 +1,7 @@
 import { Checkbox, FormControl, ListItemText, MenuItem, Select } from "@mui/material";
 import { Filter } from "../src/sortingAndFiltering";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
-import styles from './SortAndFilterDropdowns.module.css'
+import styles from '../styles/SortAndFilterDropdowns.module.css'
 
 // Customising the table styling using ThemeProvider
 const theme = createTheme({
