+++ components/SummaryComponents/UrgentRepos.tsx
@@ -1,7 +1,7 @@
 import Table from "@mui/material/Table";
 import TableCell from "@mui/material/TableCell";
 import TableRow from "@mui/material/TableRow";
-import styles from "../../components/summary_components/UrgentRepos.module.css";
+import styles from "../../styles/UrgentRepos.module.css";
 
 export default function UrgentRepos() {
   return (
