+++ styles/CondensedSummary.module.css
