+++ styles/DependenciesContainer.module.css
